add a new unsurprised condition to make combat log clear
