+++ SolastaCommunityExpansion/Translations-en.txt
@@ -614,3 +614,5 @@ Subclass/&BarbarianPathOfTheLightIlluminatingBurstPowerDescription	Up to three e
 Subclass/&BarbarianPathOfTheLightIlluminatingBurstPowerTitle	Illuminating Burst
 Feedback/&AdditionalDamageBarbarianPathOfTheLightIlluminatingStrikeLine	Illuminating Strike deals extra damage to the target!
 Feedback/&AdditionalDamageBarbarianPathOfTheLightIlluminatingStrikeFormat Illuminating Strike!
+Rules/&ConditionUnsurprisedDescription	was surprised but succeed on a perception check.
+Rules/&ConditionUnsurprisedTitle	no longer surprised
